small fix, +tests

+++ integrator_pt_scene.cpp
@@ -731,7 +731,7 @@ bool Integrator::LoadScene(const char* a_scenePath, const char* a_sncDir)
       uint32_t offset = m_wavelengths.size();
       std::copy(spec.wavelengths.begin(), spec.wavelengths.end(), std::back_inserter(m_wavelengths));
       std::copy(spec.values.begin(), spec.values.end(), std::back_inserter(m_spec_values));
-      m_spec_offset_sz.push_back(uint2{offset, spec.wavelengths.size()});
+      m_spec_offset_sz.push_back(uint2{offset, static_cast<uint32_t>(spec.wavelengths.size())});
       
       // we expect dense, sorted ids for now
       //assert(m_spectra[spec_id].id == spec_id);
@@ -748,7 +748,7 @@ bool Integrator::LoadScene(const char* a_scenePath, const char* a_sncDir)
       uint32_t offset = m_wavelengths.size();
       std::copy(uniform1.wavelengths.begin(), uniform1.wavelengths.end(), std::back_inserter(m_wavelengths));
       std::copy(uniform1.values.begin(), uniform1.values.end(), std::back_inserter(m_spec_values));
-      m_spec_offset_sz.push_back(uint2{offset, uniform1.wavelengths.size()});
+      m_spec_offset_sz.push_back(uint2{offset, static_cast<uint32_t>(uniform1.wavelengths.size())});
     }
   }
